Move lua db to helpers
